remove error module
