--- crates/ibc/src/core/ics24_host/error.rs
@@ -1,25 +1,0 @@
-use crate::prelude::*;
-use displaydoc::Display;
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize))]
-#[derive(Debug, Display)]
-pub enum IdentifierError {
-    /// identifier `{id}` cannot contain separator '/'
-    ContainSeparator { id: String },
-    /// identifier `{id}` has invalid length `{length}` must be between `{min}`-`{max}` characters
-    InvalidLength {
-        id: String,
-        length: usize,
-        min: usize,
-        max: usize,
-    },
-    /// identifier `{id}` must only contain alphanumeric characters or `.`, `_`, `+`, `-`, `#`, - `[`, `]`, `<`, `>`
-    InvalidCharacter { id: String },
-    /// identifier prefix `{prefix}` is invalid
-    InvalidPrefix { prefix: String },
-    /// identifier cannot be empty
-    Empty,
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for IdentifierError {}
