+++ crates/ibc/src/core/ics24_host/mod.rs
@@ -3,7 +3,6 @@
 
 pub use path::{ClientUpgradePath, Path, IBC_QUERY_PATH, SDK_UPGRADE_QUERY_PATH};
 
-pub mod error;
 pub mod identifier;
 pub mod path;
 pub(crate) mod validate;
